parser.c分割
